requirements.txt
grpcio==1.27.2
APScheduler==3.6.3
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
cryptography==40.0.1
decorator==5.1.1
et-xmlfile==1.1.0
phonenumbers==8.13.10
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.0
python-telegram-bot-pagination==0.0.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.22.0
six==1.16.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.25.7

